requirements.txt
# Azure OpenAI client
openai>=1.2.0

# Data handling
pandas>=2.1.0

# Progress bar
tqdm>=4.66.1

# Machine learning metrics
scikit-learn>=1.3.0

# Plotting for confusion matrix
matplotlib>=3.8.1
seaborn>=0.12.2

# YAML configuration
PyYAML>=6.0

# Retry/backoff handling
backoff>=2.2.1
